requirements.txt
ipykernel
langchain
python-dotenv
langchain-community
pypdf
bs4
arxiv
pymupdf
fitz
wikipedia
langchain-text-splitters
sentence_transformers
langchain-openai
chromadb
langchain_huggingface
faiss-cpu
langchain_chroma
langchain-groq
beautifulsoup4
streamlit